+++ ouroboros-network-protocols/src/Ouroboros/Network/Protocol/ChainSync/Client.hs
@@ -21,14 +21,9 @@ module Ouroboros.Network.Protocol.ChainSync.Client
   , ClientStIntersect (..)
     -- * Execution as a typed protocol
   , chainSyncClientPeer
-    -- * Null chain sync client
-  , chainSyncClientNull
     -- * Utilities
   , mapChainSyncClient
   ) where
-
-import Control.Monad (forever)
-import Control.Monad.Class.MonadTimer
 
 import Network.TypedProtocol.Core
 import Network.TypedProtocol.Peer.Client
@@ -42,12 +37,6 @@ newtype ChainSyncClient header point tip m a = ChainSyncClient {
     runChainSyncClient :: m (ClientStIdle header point tip m a)
   }
 
--- | A chain sync client which never sends any message.
---
-chainSyncClientNull :: MonadTimer m => ChainSyncClient header point tip m a
-chainSyncClientNull = ChainSyncClient $ forever $ threadDelay 43200 {- one day in seconds -}
-
-{-# DEPRECATED chainSyncClientNull "Use Ouroboros.Network.NodeToClient.chainSyncPeerNull" #-}
 
 -- | In the 'StIdle' protocol state, the server does not have agency and can choose to
 -- send a request next, or a find intersection message.