+++ ouroboros-network-protocols/src/Ouroboros/Network/Protocol/LocalTxSubmission/Client.hs
@@ -21,14 +21,9 @@ module Ouroboros.Network.Protocol.LocalTxSubmission.Client
   , SubmitResult (..)
     -- * Execution as a typed protocol
   , localTxSubmissionClientPeer
-    -- * Null local tx submission client
-  , localTxSubmissionClientNull
     -- * Utilities
   , mapLocalTxSubmissionClient
   ) where
-
-import Control.Monad (forever)
-import Control.Monad.Class.MonadTimer
 
 import Network.TypedProtocol.Core
 import Network.TypedProtocol.Peer.Client
@@ -40,13 +35,6 @@ newtype LocalTxSubmissionClient tx reject m a = LocalTxSubmissionClient {
       runLocalTxSubmissionClient :: m (LocalTxClientStIdle tx reject m a)
     }
 
--- | A local tx submission client which never sends any message.
---
-localTxSubmissionClientNull :: MonadTimer m => LocalTxSubmissionClient tx reject m a
-localTxSubmissionClientNull =
-    LocalTxSubmissionClient $ forever $ threadDelay 43200 {- day in seconds -}
-
-{-# DEPRECATED localTxSubmissionClientNull "Use Ouroboros.Network.NodeToClient.localTxSubmissionPeerNull" #-}
 
 -- | The client side of the local transaction submission protocol.
 --