+++ ouroboros-network-protocols/src/Ouroboros/Network/Protocol/TxSubmission2/Client.hs
@@ -21,7 +21,6 @@ module Ouroboros.Network.Protocol.TxSubmission2.Client
   , ClientStIdle (..)
   , ClientStTxIds (..)
   , ClientStTxs (..)
-  , TxSizeInBytes
   , SingBlockingStyle (..)
   , BlockingReplyList (..)
     -- * Execution as a typed protocol