+++ ouroboros-network-framework/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Breaking changes
 
 * Addapted to `network-mux` changes in https://github.com/IntersectMBO/ouroboros-network/pull/4997
+* Removed deprecated `Ouroboros.Network.Channel.{to,from}Channel` functions.
 
 ### Non-breaking changes